Delete soroban-react, adding PairClient
