+++ packages/state/src/clients/index.ts
@@ -1,0 +1,2 @@
+export * from "./Pair.client";
+export * from "./Token.client";
