--- packages/state/src/hooks/contracts/index.tsx
@@ -1,1 +1,0 @@
-export * from "./useContractQuery";
