--- packages/state/src/providers/walletProvider/index.ts
@@ -1,1 +1,0 @@
-export { SorobanReactProvider as WalletProvider } from "./soroban/SorobanReactProvider";
