--- packages/state/src/providers/walletProvider/soroban/getDefaultConnectors.tsx
@@ -1,8 +1,0 @@
-import { Connector } from "../types";
-import { freighter } from "../../../state/wallet/freighter";
-
-export const getDefaultConnectors = (): Connector[] => {
-  const list: Connector[] = [freighter()];
-
-  return list;
-};
