--- packages/state/src/providers/walletProvider/soroban/index.tsx
@@ -1,4 +1,0 @@
-export * from "./SorobanContext";
-export * from "./SorobanReactProvider";
-export * from "../hooks/useSorobanReact";
-export * from "./getDefaultConnectors";
