+++ packages/state/src/state/wallet/actions.ts
@@ -15,50 +15,49 @@ export const createWalletActions = (
 ): WalletActions => {
   return {
     // The address of the wallet connected to the app
-    wallet: undefined,
+    wallet: {
+      address: undefined,
+      activeChain: undefined,
+      server: undefined,
+    },
 
     // This function is called when the user clicks the "Connect" button
     // in the wallet modal. It uses the Freighters SDK to get the user's
     // address and network details, and then stores them in the app state.
     connectWallet: async () => {
-      try {
-        // Get the network details from the user's wallet.
-        const networkDetails = await freighter().getNetworkDetails();
-
-        // Throw an error if the network is not supported.
-        if (
-          !allChains.find(
-            (c: any) =>
-              c.networkPassphrase === networkDetails?.networkPassphrase
-          )
-        ) {
-          const error = new Error(
-            "Your Wallet network is not supported in this app"
-          );
-          throw error;
-        }
-
-        // Get the active chain from the network details.
-        const activeChain = networkToActiveChain(networkDetails, allChains);
-
-        // Get the user's address from the wallet.
-        const address = await freighter().getPublicKey();
-
-        // Create a server object to connect to the blockchain.
-        let server =
-          networkDetails &&
-          new Server(networkDetails.networkUrl, {
-            allowHttp: networkDetails.networkUrl.startsWith("http://"),
-          });
-
-        // Update the state to store the wallet address and server.
-        setState((state: AppStore) => ({
-          ...state,
-          wallet: { address, activeChain, server },
-        }));
-      } catch (error) {
-        // Handle error
+      // Get the network details from the user's wallet.
+      const networkDetails = await freighter().getNetworkDetails();
+      console.log(4);
+      // Throw an error if the network is not supported.
+      if (
+        !allChains.find(
+          (c: any) => c.networkPassphrase === networkDetails?.networkPassphrase
+        )
+      ) {
+        const error = new Error(
+          "Your Wallet network is not supported in this app"
+        );
+        throw error;
       }
+
+      // Get the active chain from the network details.
+      const activeChain = networkToActiveChain(networkDetails, allChains);
+
+      // Get the user's address from the wallet.
+      const address = await freighter().getPublicKey();
+
+      // Create a server object to connect to the blockchain.
+      let server =
+        networkDetails &&
+        new Server(networkDetails.networkUrl, {
+          allowHttp: networkDetails.networkUrl.startsWith("http://"),
+        });
+
+      // Update the state to store the wallet address and server.
+      setState((state: AppStore) => ({
+        ...state,
+        wallet: { address, activeChain, server },
+      }));
     },
 
     // Disconnect the wallet
@@ -66,7 +65,11 @@ export const createWalletActions = (
       // Update the state
       setState((state: AppStore) => ({
         ...state,
-        wallet: undefined,
+        wallet: {
+          address: undefined,
+          activeChain: undefined,
+          server: undefined,
+        },
       }));
     },
 
@@ -91,9 +94,7 @@ export const createWalletActions = (
       );
 
       // Fetch token balance
-      const balance = await queryClient.balance(
-        getState().wallet?.address as string
-      );
+      const balance = await queryClient.balance();
 
       // Update token balance
       setState((state: AppStore) => {
