+++ packages/utils/src/constants.ts
@@ -1,0 +1,6 @@
+import { Account } from "soroban-client";
+
+export const TESTING_SOURCE: Account = new Account(
+  "GBUHRWJBXS4YAEOVDRWFW6ZC5LLF2SAOMATH4I6YOTZYHE65FQRFOKG2",
+  "1"
+);
